requirements.txt
asgiref==3.10.0
crispy-bootstrap5==2025.6
Django==5.2.8
django-crispy-forms==2.5
psycopg==3.1.18
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
djangorestframework==3.15.0
gunicorn